fix TransactionConfirmation for InstantSeal (#936)
* fix TransactionConfirmation for InstantSeal

since parity --chain dev is using its InstantSealEngine and creating Block as soon as a new Transaction is send, the subscribed event for the new BlockHeader will come too late. This results in a waiting forever for a new Block, which never comes. So we should first check whether the transaction has already been mined before subscribing for new Blocks.

* fixed format and added bind

* keep on listing only if listeners are registered

* fixed tests

* added test for instantSeal

* cleanup