Handle Revert code and throw containing message
Hi @frozeman 
I'm not an expert at all about web3, but I find out that it doesn't handle require/revert messages (https://solidity.readthedocs.io/en/v0.4.24/control-structures.html?highlight=revert#error-handling-assert-require-revert-and-exceptions).

I don't know what is a good way to do it, and I don't even know how to parse Ethereum response correctly. But I had to implement this feature for my product somehow and I've decided to share this code with you so then you can decide how to do it better.
Or maybe I can finish this if you give me some tips on where to look for the decoder and how to better organize the code.