pickup latest fsevents (v1.2.9) for Node 12 compat
fsevents is a transitive dependency by way of jest. `package-lock.json` was
doing its job and preventing a bump to fsevents v1.2.9, i.e. even though the
caret ranges for that dependency would allow for it.

Older versions of `fsevents` are not compatible with Node.js 12, so for this
commit `package-lock.json` was manually bumped (resolved fsevents was removed
from the file and then `npm install` was rerun).