implement fallback from built-in scrypt to scryptsy/scrypt packages
This commit introduces an additional fallback behavior and console warning in
the case that Node's built-in scrypt experiences memory exhaustion, as was
discovered in the 1.x test suite with the static tests (see #2938). A large `n`
parameter is the culprit, but that's not the case when using the 3rd
party (deprecated) scrypt package. The `scryptsy` package can handle a large
`n`, but performance does suffer considerably.