Make sendRequest return type consistent
The inner function sendRequest that belongs to Method.buildCall should not return anything. Returning a promise might lead to some dev reading the code to think that any promise rejection would be handled in the caller. But the fact is that all the calls to it (below) just discard anything returned. So better to clean this up.