fix(validator): generate abi name if not provided
In the case of public mappings, the ABI generated has a blank name set.
This results in blank ids for such inputs within the JSON schema. Post
1f81ff097cf548a327fb72537dbabf2126e616d0, the number of unique ids is
used to convert the JSON scheme into Zod, and hence the presence of
blank ids is a concern.

This commit generates the `abiName` if one is not provided to a value
equal to the `${level}/${index}`.

Fixes #6965.